Addition of testing as well as updates to allow bucket overriding

+++ index.js
@@ -1,21 +1,35 @@
 
 const SimpleS3DB = function( s3, bucket ) {
 
-    this.put = function( key, obj ) {
+    this.put = function( key, obj, o ) {
+        const options = o || {};
+
         return new Promise( ( resolve, reject ) => {
 
+            if ( !( options.bucket || bucket ) ) {
+                return reject( new Error( 'No bucket provided for PUT request' ) );
+            }
+
             try {
-                s3.putObject( {
-                    Bucket: bucket,
+                const s3Object = {
+                    Bucket: options.bucket || bucket,
                     Key: key,
                     Body: typeof obj === 'object' ? JSON.stringify( obj ) : obj
-                }, ( err, data ) => {
+                };
+
+                Object.keys( options ).forEach( currentOptionKey => {
+                    if ( currentOptionKey !== 'bucket' ) {
+                        s3Object[ currentOptionKey ] = options[ currentOptionKey ];
+                    }
+                } );
+
+                s3.putObject( s3Object, ( err, data ) => {
                     if ( err ) {
                         return reject( err );
                     }
 
-                    resolve( data );
-                });
+                    resolve( obj );
+                } );
             } catch( e ) {
                 reject( e );
             }
@@ -23,12 +37,16 @@ const SimpleS3DB = function( s3, bucket ) {
         } );
     };
 
-    this.get = function( key ) {
+    this.get = function( key, bucketOverride ) {
         return new Promise( ( resolve, reject ) => {
+
+            if ( !( bucketOverride || bucket ) ) {
+                return reject( new Error( 'No bucket provided for Get request' ) );
+            }
 
             try {
                 s3.getObject( {
-                    Bucket: bucket,
+                    Bucket: bucketOverride || bucket,
                     Key: key
                 }, ( err, data ) => {
                     if ( err ) {
@@ -58,14 +76,18 @@ const SimpleS3DB = function( s3, bucket ) {
         } );
     };
 
-    this.delete = function( key ) {
+    this.delete = function( key, bucketOverride ) {
         return new Promise( ( resolve, reject ) => {
+
+            if ( !( bucketOverride || bucket ) ) {
+                return reject( new Error( 'No bucket provided for DELETE request' ) );
+            }
 
             if ( Array.isArray( key ) ) {
 
                 try {
                     s3.deleteObjects( {
-                        Bucket: bucket,
+                        Bucket: bucketOverride || bucket,
                         Delete: {
                             Objects: key.map( currentKey => {
                                 return { Key: currentKey }
@@ -86,7 +108,7 @@ const SimpleS3DB = function( s3, bucket ) {
 
                 try {
                     s3.deleteObject({
-                        Bucket: bucket,
+                        Bucket: bucketOverride || bucket,
                         Key: key
                     }, (err, data) => {
                         if (err) {
@@ -104,12 +126,16 @@ const SimpleS3DB = function( s3, bucket ) {
         } );
     };
 
-    this.list = function( folder ) {
+    this.list = function( folder, bucketOverride ) {
 
         return new Promise( ( resolve, reject ) => {
 
+            if ( !( bucketOverride || bucket ) ) {
+                return reject( new Error( 'No bucket provided for LIST request' ) );
+            }
+
             s3.listObjectsV2( {
-                Bucket: bucket,
+                Bucket: bucketOverride || bucket,
                 Delimiter: "/",
                 Prefix: folder.endsWith( '/' ) ? folder : folder + '/'
             }, ( err, data ) => {
